% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class1.R
\name{+.class1}
\alias{+.class1}
\alias{+}
\title{Add objects of type class 1}
\usage{
+x
x + y
}
\arguments{
\item{x}{an object of class "class1"}

\item{y}{an object of class "class1"}
}
\value{
an object of class "class1"
}
\description{
Add objects of type class 1
}
