% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class2.R
\name{+.class2}
\alias{+.class2}
\alias{+}
\title{Add objects of type class 1}
\usage{
+x
x + y
}
\arguments{
\item{x}{an object of class "class2"}

\item{y}{an object of class "class2"}
}
\value{
an object of class "class2"
}
\description{
Add objects of type class 1
}
